Rename experiment home page to experiment page to avoid name clash
